Detect the best generic level before real uarch
The detection algorithm is changed to detect first
the best generic level for the current architecture,
and then try to detect the real microarchitecture.

Candidates for the real microarchitecture must
support all the features of the best generic
architecture. If no real uarch is detected, then
the best generic is returned.