Detect the best generic level before the real microarchitecture (#53)
In some cases our compatibility algorithm can identify chips in strange ways.
We identify a host by finding the "nearest" compatible uarch -- i.e. the uarch 
with the maximal subset of features w.r.t the host.  In some cases, that can have
unexpected results. e.g., a `haswell` can be identified as `nehalem` if the `aes`
instructions are missing, because that is the closest ancestor without `aes`. 
However, we would much rather it be identified as some architecture with `avx2`
instructions.

`x86_64` levels like `x86_64_v2`, `x86_64_v3` are generic uarches that focus
on vector instructions, and they provide the baselines we need for this.  We can
safely detect a `haswell` as `x86_64_v3`, which has `avx2`.  This no longer 
follows our "maximal subset" rule, but it gives users an architecture id they
more likely want.

This PR changs the detection algorithm to first detect the best generic level
for the current architecture (`v4`, `v3`, `v2`, etc.), THEN try to find a maximal
subset of features starting from the best generic. 

Candidates for the real microarchitecture must support all the features of the
best generic architecture. If no real uarch is detected, then the best generic is
returned.