Add support for CPU detection on Windows (#150)
This PR builds on #129 and #149 to add support for Windows. Features detection is done using `cpuid`. The `cpuid` calls that we need to detect features are in https://github.com/archspec/archspec-json/pull/86. We also forked [archspec/cpuid](https://github.com/archspec/cpuid) and use it here as a submodule.

Modifications:
- [x] Add a new method to detect CPU features and vendor using `cpuid`
- [x] Add unit tests for the new method, including Windows
- [x] update archspec-json

---------

Co-authored-by: Isuru Fernando <isuruf@gmail.com>