Using cpuid field for AARCH64 microarchitectures
This patch introduces the use of cpuid for
microarchitecture detection for aarch64. An object
of type AARCH64_Microarchitecture has been created
to contain the cpuid field which might be supplied
from the Json file of microarchitectures (from the
json submodule). The same field is then used in
the compatibiltiy_checks function.